MinHash.jl

Efficient minhashing in Julia

MinHash.jl offers generic, efficient MinHash sketching, and functions to efficiently compute the number of shared minhashes between sketches. This package is envisioned to be used as a dependency for other Julia packages that needs minhashing.

The API is very simple. Below is a short demonstration - read the docstrings below for more info.

julia> using MinHash

julia> hasher = MinHasher(100) # Same as MinHasher{hash}(100), keep 100 hashes
MinHasher{hash}:
 hashes:  0 / 100
 maxhash: < uninitialized >
 hashset: HashSet(0 / 512)

julia> update!(hasher, 'a':'z') # update with any iterable
MinHasher{hash}:
 hashes:  26 / 100
 maxhash: < uninitialized >
 hashset: HashSet(26 / 512)

julia> update!(hasher, 1:100) # only 100 hashes are kept
MinHasher{hash}:
 hashes:  100 / 100
 maxhash: 0xd0196f660622f483
 hashset: HashSet(126 / 512)

julia> sk = MinHashSketch(hasher) # can be used on noninitialized hasher
MinHashSketch:
 hashes:  100 / 100
 maxhash: 0xd0196f660622f483

julia sk2 = sketch(1:1000, 100) # quickly keep 100 hashes of 1:1000
MinHashSketch:
 hashes:  100 / 100
 maxhash: 0xfff598f76319355e

julia> intersectionlength(sk, sk2) # calc number of hashes in sk and sk2
80

MinHash.HashSet — Type

HashSet(N::Integer)

Construct a HashSet with at least N slots. A HashSet contains UInt hash values and is optimized for quick inserts and membership test.

A HashSet is a completely unsafe type internal type, not to be used unless you know what you are doing.

MinHash.MinHashSketch — Type

MinHashSketch

Holds the N smallest hashes of an iterable. Construct from a MinHasher.

Examples

julia> x = MinHashSketch(update!(MinHasher(10), 1:1000))
MinHashSketch:
 hashes:  10 / 10
 maxhash: 0x0214ce7a1c004d40

julia> length(x)
10

MinHash.MinHasher — Type

MinHasher{F}

A type used to minhash any iterable object with function F.

See also: update!, sketch

MinHash.intersectionlength — Method

intersectionlength(v::AbstractVector{MinHashSketch})

Return a length(v) square Matrix{Int} where the [i, j] where i > j'th cell contain the number of hashes in both v[i] and v[j].

Examples

julia> v = [sketch((1:1000) .+ (300*i), 100) for i in 1:4];

julia> intersectionlength(v)
4×4 Array{Int64,2}:
  0   0   0  0
 64   0   0  0
 27  58   0  0
  8  34  73  0

MinHash.intersectionlength — Method

intersectionlength(y::MinHashSketch, y::MinHashSketch)

Count number of hashes in both x and y more effectively than using ordinary set operations.

Examples

julia> x = sketch(1:10000, 100); y = sketch(5000:15000, 100);

julia> intersectionlength(x, y)
49

MinHash.sketch — Method

sketch([F=hash], it, s::Integer)

Hash every element of iterable it using function F, and return a MinHashSketch containing at most the s smallest hashes.

Examples:

julia> sketch("ACGDEFG", 3)
MinHashSketch:
 hashes:  3 / 3
 maxhash: 0x3e1b023d3c92ff8f

See also: update!

MinHash.update! — Method

update!(s::MinHasher, it)

Add hashes of all elements of iterable it to MinHasher s.

See also: sketch